Change transport retrieval to support multiple connection.
Originally it was intended for one transport per resource, but for
vShield we need multiple transport per resource and the ability to
discover the vShield associated vCenter system without the user
specifying the connection. This update allows discovery of the
associated vCenter system through vShield API calls, but use the
connection info from the manifest (since it's not available in vShield).