Allow numbers for mtu
Currently, the implementation of the esx_portgroup
type does not support numbers as arguments b/c they
call a string method strip on all values.

This patch ensures that all values are converted into
strings before strip is called on them.

It also updates the unit tests to ensure that numbers
can be passed in as arguments to this type.

I had a few issues running the tests:

- many of the tests from this file seemed to fail
b/c enabled with a capital letter 'E' is not supported.

- I encountered additional test failures b/c it called
the method stub instead of stubs (which I believe is
related to running the tests without mocha enabled
for mocking)