Explicitly pass transports for esx_service
Previously, the esx_service resources were not
being passed in their transports explicitly,
even though they require the transport to be
specified.

They were being set when the asm::host was
specified as a resource b/c transport is a metaparameter
which, when specified for a container is forwarded to
resources in that container.

For some reason, this behavior worked for resources
specified in a manifest, and not resources passed
through our yaml node terminus (I have not been
able to track down why this is the case).