allow esx resources to work without vcenter
Many esx resources searched for the esx host to configure under a
specific datacenter. This prevents the esx host itself from being used
as the transport because the host does not exist in a datacenter.

This patch falls back to searching for the host by ip or dns name
alone if no path is specified for the host.