Create VM on local storage
added a param in vc_vm resource
newparam(:skip_local_datastore) do
newvalues(:true,:false)
defaultto(:false)
end

When a VM is created on the cluster and the value for this flag is set to true, we will skip local datastore from the candidate datastore. In the ASM manifest asm::vm::vcenter added a new parameter "skip_local_datastore" with default value as "true". This will ensure that we will skip the local datastore from the candidate datastore list.