ASM-2873 RHEL VM does not get the network connectivity
The recent changes done for avoiding the VM reset during the VM creation and retry has introduced this error.
We were trying to change the port-group mapping without deleting and re-adding a new network interface. But the internal input spec generated in the puppet module was trying to remove the existing network interface for just changing the port-group mapping.

Checks are added to the puppet module to take care of the scenario where the count of existing network and network getting updated are the same then avoid the deletion of the network and just change the port-group mapping.