Update default.rb
Changed the search mechanism for a host to use the DNS name as the name is passed in and used by other functions when matching the host. Without this change dvswitch registration to fails because the host moRef cannot be returned. If an IP address is passed in to get this function to return correctly then the host match in the list match in mo_ref_by_name fails.