ASM-6057 Management VNIC is not getting mapped correctly for VDS deployments
iSCSI initiator mapping puppet resource is using esxcli commands for mapping vmkernel with software adapters. In case vmkernel is already binding to the software adapter, error was generated.
so far we were having one to mapping between iSCSI VMK and storage adapter thats why this error was not surfaced

with software initiator we need to map multiple vmkernel to same storage adapter.