ASM-6600 [puppet] Redirect vSAN traces to an external datastore
As per the story, we need to redirect VSAN traces to external datastore.
In the puppet module, we are passed a volume name to set for all hosts,
and we simply apply that.

The business logic of determining the right volume name is done in
asm-deployer. We set vsan traces to this datastore corresponding to this
volume name using the relevant esxcli API