Add timing information when rebooting host during VIB updates
Seems there are cases where host reboot may take longer than
anticipated. We need to add information in logs to know how long
it took to connect for debugging purpose. Also we add a log entry if
we find the host reboot timeout happened and we were still not able
to have connected host.