Attempt to install as many VIBs as possible
Found out that if there are invalid VIB binaries sent, we can end up in
an 'NilClass' exception from rbvmomi when installing the VIB. Instead of
stopping at error vib, we want to install as many VIBs as possible, and
log any failures.

In the end we altready have logic to fail the update if count of installed
VIBs does not equal count of "to install" VIBs.

Note: Also we change Puppet.error to Puppet.err since Puppet.error is not
defined.  Perhaps we should have rubocop to catch this.