ASM-7658 Support installing VIBs in a zip for use cases with dependencies
The esxcli API to install vibs already supports passing an offline bundle
which is essentially a ZIP with vibs. All dependent VIBs are part of this
bundle and seems to work if installed as single batch.

In addition to above changes, we also do a peformance optimization to not
query a source VIB for vib_id as it is already passed via instance_id to
this puppet module.

Also did a minor refactor to move some of the pieces in exists? method
into its own methods.