ASM-7676 Failed to unmount NFS share after rebooting host
For software updates on VIBs, if some of the VIBs are not installed, then
we may end up scheduling NFS unmount after rebooting the ESX host which
seems to not work.  We can see strange error message due to this:
RuntimeError:esxcli not supported through VC before 5.0.0

The fix is to unmount before reboot operation.

Also in this PR, we remove the unnecessary begin/rescue for entire
install operation, which is only need for install_vib.