ASM-7784 Ensure host is in maintenance mode before starting updates
Background of the problem:
Sometimes, if there is a pre-exisiting firmware updates before a software
update, after rebooting the host may lose its maintenance mode state.
This can cause issue "Cannot perform the operation in current state"
on invoking wait_for_completion on a reboot action on the host. Also NFS
datastores to the appliance do not get unmounted as another side-effect.

Solution:
To solve the issue of having unmounted datastore left behind as well as
ensure reboot host succeeds, we need to ensure maintenance mode is set
on the host before we kick off software updates.