ASM-7861  Ensure no VSAN action for maintenance mode during server updates
Seems if we do not have this flag, VSAN hosts can fail to enter maintenance
mode because in absence of VSAN action not supplied then VSAN datastore
integrity and accessibility test takes a lot of time and leads to timeout.

If host is not a VSAN host, this flag will not impact maintenance mode.