ASM-7987 fix iscsi binding provider call to esxcli
The vcenter puppet module was failing to bind the iscsi adapters
correctly due to the lack of special characters handling before
making the system call.

In order to resolve this, ASM::Util.run_command method was chosen
instead of making system call directly from the provider script.

Additionally, the password was set as an environment variable for
esxcli (VI_PASSWORD) instead of being passed directly. This should
stop the process list from showing the password.

UPDATE: Changed the parameter style of the server input for the
esxcli to match the rest. Enabled additional validation and logg-
ing. Moved the logging locations around and added safeguard for
rescued process.