ASM-8072 Increase vcenter discovery timeout to 15 minutes
In scenarios with large vCenters (~30 hosts, 1k+ VMs) the discovery
script is seen to take about 10 minutes. This changes simply bumps the
default timeout up from 4 minutes to 15 minutes to account for
that. In the future we may want to see if we can improve the speed of
discovery perhaps by inventorying less data on individual hosts.