ScaleIO VM Support
While passing the specific datastore name for VM creation, operation was failing in the debug message.
Also in the return statement we were just sending the name of the datastore (old convention) but with the new code structure, datastore object information was required.
Validated the code for VM creation on specific host in the cluster.