VXFM-3435 Wrong disk mapped on CO deployment
Included parameter disk model and disk id for searching the disk that needs to be mapped and mounted as a datastore. In case there are multile local disks, then all of them becomes cadidate to be mounted as the datastore. Now we we will check for model, target id and lun id before mounting the volume as a datastore