VXFM-3187 Management vmnic migration failure
For VMNIC creation, there is a possibility of multiple NICs getting created in parallel. VMNIC ID is retreived based on the previously existing VMKNIC. In case two operations gets the same VMKNIC, then one of the operation will fail.
Added a delay and retry for VMNIC creation to avoid failure during initiatial VMKNIC create operation.

Possibility of this error is remote but this generally leads to the failure of whole puppet agent run.